requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1                  # 4.1+ 에서 passlib 호환 깨짐
python-multipart==0.0.9

# Async / Queue
celery==5.3.6
redis==5.0.4

# AI providers (필요한 것만 주석 해제)
openai==1.30.1
anthropic==0.26.1
# google-generativeai==0.5.4

# Validation / Settings
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# HTTP
httpx==0.27.0
tenacity==8.3.0   # retry 로직

# Utils
python-dateutil==2.9.0
structlog==24.2.0  # structured logging

# Dev / Test
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0      # TestClient용
faker==25.0.0
